Criteria Builder / Dynamic Query
added a new repository and service that holds a method, for now, used for filtering a search based on params

needed this because I tried a query annotation on top of my method that made some joins. Worked, but the problem showed up when IN was used. It seems that lists cannot be passed anymore into the queries, older spring boot versions supported it...

some new dependencies in pom

cr: ChirilaT